x/gitopia/types/message_update_repository_backup_ref_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/gitopia/gitopia/testutil/sample"
	"github.com/stretchr/testify/require"
)

func TestMsgUpdateRepositoryBackupRef_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgUpdateRepositoryBackupRef
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgUpdateRepositoryBackupRef{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgUpdateRepositoryBackupRef{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
